Price ok
Завершение работы над страницей прайса услуг по ремонту смартфонов

+++ lib/components/constants.dart
@@ -12,7 +12,7 @@ import 'package:flutter/material.dart';
   const double kDefaultRadius = 30.0;
   const double kCardDefaultHeigth = 200 + kDefaultPadding * 2;
   const double kImagePrice = 20.0;
-  const int kMobilSize = 5;
-  const int kTabletSize = 6;
-  const int kDesktopSize = 7;
+  const double kMobilSize = 5;
+  const double kTabletSize = 6;
+  const double kDesktopSize = 7;
